+++ docs/source/changelog.rst
@@ -4,7 +4,9 @@ Change Log
 
 **Pyro 4.51**
 
-...
+- added PYROMETA magic URI protocol, to look up an object with the given metadata tags (yellow-page lookup rather than by name)
+  @todo docs
+- fixed unlikely but possible crash in logging statement when client disconnects from multiplex server
 
 
 **Pyro 4.50**